@@ -27,7 +27,7 @@ abstract class NewExcelFile extends File {
     public function handleExport()
     {
         return $this->handle( 
-            get_class($this) 
+            static::class 
         );
     }
 
